PythonFullStackProject/requirements.txt
streamlit>=1.29
supabase>=2.0.2
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0